Fixed an error in Parser, more needs to be fixed. Changed format of README file so git will recognise it.
Fixed an error in Parser, more needs to be fixed. Changed format of README file so git will recognise it.

Fixed an error in Parser, more needs to be fixed. Changed format of README file so git will recognise it.

Fixed an error in Parser, more needs to be fixed. Changed format of README file so git will recognise it.