Began implementation of Goal classes. The types of goals extend the Goal class (inheritance). Added constructor and getters to goal class and the createGoal(), addGoal(), removeGoal() and checkGoals() methods to the Goals class.
Began implementation of Goal classes. The types of goals extend the Goal class (inheritance). Added constructor and getters to goal class and the createGoal(), addGoal(), removeGoal() and checkGoals() methods to the Goals class.

Began implementation of Goal classes. The types of goals extend the Goal class (inheritance). Added constructor and getters to goal class and the createGoal(), addGoal(), removeGoal() and checkGoals() methods to the Goals class.

Began implementation of Goal classes. The types of goals extend the Goal class (inheritance). Added constructor and getters to goal class and the createGoal(), addGoal(), removeGoal() and checkGoals() methods to the Goals class.