changed constructors of the individual goal classes so that the start date, target date and goal type are only prompted for in the Goal class. These classes also now use the DateTime class to set the start and target times / dates.
changed constructors of the individual goal classes so that the start date, target date and goal type are only prompted for in the Goal class. These classes also now use the DateTime class to set the start and target times / dates.

changed constructors of the individual goal classes so that the start date, target date and goal type are only prompted for in the Goal class. These classes also now use the DateTime class to set the start and target times / dates.

changed constructors of the individual goal classes so that the start date, target date and goal type are only prompted for in the Goal class. These classes also now use the DateTime class to set the start and target times / dates.