Updated DateTime class to accept values in order of magnitude when constructing, also changed the error messages to be more detailed. Updated DateTimeTest to match changes in DateTime.
Updated DateTime class to accept values in order of magnitude when constructing, also changed the error messages to be more detailed. Updated DateTimeTest to match changes in DateTime.

Updated DateTime class to accept values in order of magnitude when constructing, also changed the error messages to be more detailed. Updated DateTimeTest to match changes in DateTime.

Updated DateTime class to accept values in order of magnitude when constructing, also changed the error messages to be more detailed. Updated DateTimeTest to match changes in DateTime.