Got rid of Velocity goal (not appropriate for target users). Wrote review progress functions for each goal type. Added constrctor to Goals class so that the userProfile object is passed through. Changed checkGoal func to use a switch statement instead on many if-else's.
Got rid of Velocity goal (not appropriate for target users). Wrote review progress functions for each goal type. Added constrctor to Goals class so that the userProfile object is passed through. Changed checkGoal func to use a switch statement instead on many if-else's.

Got rid of Velocity goal (not appropriate for target users). Wrote review progress functions for each goal type. Added constrctor to Goals class so that the userProfile object is passed through. Changed checkGoal func to use a switch statement instead on many if-else's.

Got rid of Velocity goal (not appropriate for target users). Wrote review progress functions for each goal type. Added constrctor to Goals class so that the userProfile object is passed through. Changed checkGoal func to use a switch statement instead on many if-else's.