Modified table in Data Upload Screen so that it can edit values and input entries. Code is messy but works so will require a refactor
Modified table in Data Upload Screen so that it can edit values and input entries. Code is messy but works so will require a refactor

Modified table in Data Upload Screen so that it can edit values and input entries. Code is messy but works so will require a refactor

Modified table in Data Upload Screen so that it can edit values and input entries. Code is messy but works so will require a refactor