Added errors for when the incorrect data is entered. Now goes to the login scene once the profile is created.
Added errors for when the incorrect data is entered.
Now goes to the login scene once the profile is created.

Added errors for when the incorrect data is entered.
Now goes to the login scene once the profile is created.

Added errors for when the incorrect data is entered.
Now goes to the login scene once the profile is created.