Implemented the main Screen and merged all screens together, still have to do more on it to get it done properly.
Implemented the main Screen and merged all screens together, still have to do more on it to get it done properly.

Implemented the main Screen and merged all screens together, still have to do more on it to get it done properly.

Implemented the main Screen and merged all screens together, still have to do more on it to get it done properly.