Changes to GUI controller to link all the screens together, changed the size of the screens and still need to link goals and view activities screen.
Changes to GUI controller to link all the screens together, changed the size of the screens and still need to link goals and view activities screen.

Changes to GUI controller to link all the screens together, changed the size of the screens and still need to link goals and view activities screen.

Changes to GUI controller to link all the screens together, changed the size of the screens and still need to link goals and view activities screen.