Updated upload data screen to handle exceptions for when activities are trying to be uploaded that overlap with existing activities. Whilst doing this i found a very hard to find bug in which the parser does not reset the line count to 0 after reading a file (took me ages). Also updated Activity class and User profile to throw exceptions. Also added 2 Exception classes for these cases.
Updated upload data screen to handle exceptions for when activities are trying to be uploaded that overlap with existing activities. Whilst doing this i found a very hard to find bug in which the parser does not reset the line count to 0 after reading a file (took me ages). Also updated Activity class and User profile to throw exceptions. Also added 2 Exception classes for these cases.