Refactored createPopUp method in uploadData to be in GUIController so it can be used by all other screens.
Tidied up code style in CreateProfile screen, refactored invalid input to be handled using try catch blocks instead of booleans.

Refactored some methods to be neater and commented properly.

Fixed issue where after a user is created the values remain in the TextFields and ComboBox's etc implemented this for when the back button is pushed also.

Fixed Issue where error labels are too short and full content cannot be seen in CreateProfile screen and properly formatted the width and spacing fields of all elements to fit the containers they are in