Worked extensively on the GUI. now have a working custom title bar, a new colour theme which has been implemented on login, profile, and createprofile screens.
I have also created css files for all screens.

This has somewhat fixed the god class issue as now a lot of the functionality is shared by the title controller.

have picked up a couple of bugs i caused by refactoring it this way and i'm assuming there will be a few more.